+++ app/client/src/components/shared/WaterbodyInfo.tsx
@@ -877,9 +877,8 @@ function WaterbodyInfo({
           {
             label: (
               <>
-                Significant{' '}
-                <GlossaryTerm term="Effluent">Effluent</GlossaryTerm> Violation
-                within the last 3 years
+                Has a Significant Noncompliance Status for{' '}
+                <GlossaryTerm term="Effluent">Effluent</GlossaryTerm> Violations
               </>
             ),
             value: hasEffluentViolations ? 'Yes' : 'No',
@@ -1067,28 +1066,6 @@ function WaterbodyInfo({
         'Public Access',
         convertDomainCode(fields, 'Pub_Access', attributes.Pub_Access),
       )}
-    </>
-  );
-
-  // jsx
-  const ejscreenContent = (
-    <>
-      {labelValue('Demographic Index Percentage', attributes.T_VULEOPCT)}
-
-      {labelValue('Percent Minority', attributes.T_MINORPCT)}
-
-      {labelValue('Percent Low Income', attributes.T_LWINCPCT)}
-
-      {labelValue(
-        'Percent Less Than High School Education',
-        attributes.T_LESHSPCT,
-      )}
-
-      {labelValue('Percent Linguistically Isolated', attributes.T_LNGISPCT)}
-
-      {labelValue('Percent Individuals Under 5', attributes.T_UNDR5PCT)}
-
-      {labelValue('Percent Individuals Over 64', attributes.T_OVR64PCT)}
     </>
   );
 
@@ -1432,7 +1409,6 @@ function WaterbodyInfo({
   if (type === 'State Watershed Health Index') content = wsioContent;
   if (type === 'Alaska Native Village') content = alaskaNativeVillageContent;
   if (type === 'Protected Areas') content = protectedAreaContent;
-  if (type === 'Demographic Indicators') content = ejscreenContent;
   if (type === 'Pollutant Storage Tank') content = storageTankContent;
   if (type === 'Combined Sewer Overflow') content = sewerOverflowsContent();
   if (type === 'Wells') content = wellsContent;
@@ -1508,9 +1484,6 @@ function MapPopup({
     if (!type || typesToSkip.includes(type)) return null;
 
     let title: string | ReactNode = type;
-    if (type === 'Demographic Indicators') {
-      title = `${type} - ${feature.layer.title}`;
-    }
     if (type === 'Restoration Plans') {
       title = 'Restoration Plans for this Waterbody';
     }